\name{BlumanStats-package}
\alias{BlumanStats-package}
\alias{BlumanStats}
\docType{package}
\title{
\packageTitle{BlumanStats}
}
\description{
\packageDescription{BlumanStats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BlumanStats}
\packageIndices{BlumanStats}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BlumanStats}

Maintainer: \packageMaintainer{BlumanStats}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
